+++ changelog.md
@@ -2,6 +2,9 @@
 
 ## master
 
+- Bootstrap: Add update call to update the bootstrap server.
+- Add update api call to update the bootstrap server.
+
 ## v0.2.0
 
 - Add status information about the M3s services.